requirements.txt
python-telegram-bot==13.7
schedule==1.1.0
pytz==2023.3
python-dotenv==1.0.0
urllib3==1.26.15
six==1.16.0 